libresource: move job-exec's rset.[ch] to new library
Problem: The rset class in the job-exec module may
be generally useful, for example in the job shell
implementation.

Create a new internal library called libresource,
and include it in libflux-internal.la.  Move the
rset sources to libresource and adjust rset.h
include path in job-exec.c.
