+++ src/common/libresource/rset.c
