+++ src/common/libresource/rset.h
