+++ src/common/libresource/test/rset.c
